#19 [feat] 하루 마무리 알림을 OptionViewModel에서 처리
