+++ app/src/main/java/com/gojol/notto/ui/option/DayAlarmManager.kt
@@ -1,0 +1,8 @@
+package com.gojol.notto.ui.option
+
+interface DayAlarmManager {
+
+    fun setAlarm()
+
+    fun cancelAlarm()
+}
