+++ app/src/main/java/com/gojol/notto/ui/option/DayNotificationReceiver.kt
@@ -1,10 +1,17 @@
 package com.gojol.notto.ui.option
 
+import android.app.NotificationChannel
 import android.app.NotificationManager
+import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
+import android.os.Build
+import androidx.core.app.NotificationCompat
+import androidx.core.app.NotificationManagerCompat
 import androidx.core.content.getSystemService
+import com.gojol.notto.R
+import com.gojol.notto.ui.MainActivity
 
 class DayNotificationReceiver : BroadcastReceiver() {
 
@@ -12,8 +19,50 @@ class DayNotificationReceiver : BroadcastReceiver() {
 
     override fun onReceive(context: Context, intent: Intent) {
         notificationManager = context.getSystemService()
-        DayNotificationManager.createNotificationChannel(context)
-        DayNotificationManager.showNotification(context)
+        createNotificationChannel(context)
+        showNotification(context)
+    }
+
+    private fun createNotificationChannel(context: Context) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            val notificationChannel = NotificationChannel(
+                CHANNEL_ID,
+                CHANNEL_NAME,
+                NotificationManager.IMPORTANCE_DEFAULT
+            ).apply {
+                description = context.getString(R.string.option_notification_description)
+            }
+
+            with(NotificationManagerCompat.from(context)) {
+                createNotificationChannel(notificationChannel)
+            }
+        }
+    }
+
+    private fun showNotification(context: Context) {
+        val intent = Intent(context, MainActivity::class.java)
+        val pendingIntent = PendingIntent.getActivity(
+            context,
+            NOTIFICATION_ID,
+            intent,
+            PendingIntent.FLAG_UPDATE_CURRENT
+        )
+
+        val notificationBuilder = NotificationCompat.Builder(
+            context,
+            CHANNEL_ID
+        )
+            .setSmallIcon(R.mipmap.ic_notto_launcher_foreground)
+            .setContentTitle(context.getString(R.string.option_notification_title))
+            .setContentText(context.getString(R.string.option_notification_content))
+            .setContentIntent(pendingIntent)
+            .setPriority(NotificationCompat.PRIORITY_DEFAULT)
+            .setAutoCancel(true)
+            .setDefaults(NotificationCompat.DEFAULT_ALL)
+
+        with(NotificationManagerCompat.from(context)) {
+            notify(NOTIFICATION_ID, notificationBuilder.build())
+        }
     }
 
     companion object {
