+++ app/src/main/java/com/gojol/notto/ui/option/OptionViewModel.kt
@@ -10,7 +10,8 @@ import javax.inject.Inject
 
 @HiltViewModel
 class OptionViewModel @Inject constructor(
-    private val optionRepository: OptionRepository
+    private val optionRepository: OptionRepository,
+    private val dayAlarmManager: DayAlarmManager
 ) : ViewModel() {
 
     val isPushChecked = MutableLiveData<Boolean>()
@@ -25,6 +26,8 @@ class OptionViewModel @Inject constructor(
 
     fun updateIsPushChecked(isPushChecked: Boolean) {
         optionRepository.saveIsPushNotificationChecked(isPushChecked)
+        if (isPushChecked) dayAlarmManager.setAlarm()
+        else dayAlarmManager.cancelAlarm()
     }
 
     fun updateIsNavigateToLicenseClicked() {
